stavok/gis/pywebio-docker/requirements.txt
flask
pygeodesy
folium
pywebio
psutil
pyyaml